==> 01_Data Import and Preprocessing.ipynb <==
# The Data

IMDB Provides Several Files with varied information for Movies, TV Shows, Made for TV Movies, etc.

- Overview/Data Dictionary: https://www.imdb.com/interfaces/
- Downloads page: https://datasets.imdbws.com/

I will be focusing on the following files:

- title.basics.tsv.gz
- title.ratings.tsv.gz
- title.akas.tsv.gz

My stakeholder only wants me to include information for movies based on the following specifications:

- Exclude any movie with missing values for genre or runtime
- Include only full-length movies (titleType = "movie").
- Include only fictional movies (not from documentary genre)
- Include only movies that were released 2000 - 2021 (include 2000 and 2021)
- Include only movies that were released in the United States


Each dataset is contained in a gzipped, tab-separated-values (TSV) formatted file in the UTF-8 character set.

import pandas as pd
import numpy as np

basics_url = 'https://datasets.imdbws.com/title.basics.tsv.gz'
ratings_url = 'https://datasets.imdbws.com/title.ratings.tsv.gz'
akas_url = 'https://datasets.imdbws.com/title.akas.tsv.gz'

basics = pd.read_csv(basics_url,sep='\t', low_memory=False)

ratings = pd.read_csv(ratings_url,sep='\t', low_memory=False)

akas = pd.read_csv(akas_url,sep='\t', low_memory=False)

**Changing null value encoding from `\N` to `np.nan`**

basics = basics.replace({'\\N':np.nan})
ratings = ratings.replace({'\\N':np.nan}) 
akas = akas.replace({'\\N':np.nan}) 

**Eliminate movies that are null for `runtimeMinutes` `genres`, and `startYear`**

basics = basics.dropna(subset = ['runtimeMinutes', 'genres', 'startYear'])

**Include only movies that were released 2000 - 2021 (include 2000 and 2021)**

basics['startYear'] = basics['startYear'].astype(int)

basics = basics.loc[(basics['startYear'] >= 2000) & (basics['startYear'] <= 2021)]

**Include only full-length movies (titleType = "movie").**

basics = basics.loc[basics['titleType'] == 'movie']

**Include only fictional movies (not from documentary genre)**

is_documentary = basics['genres'].str.contains('documentary',case=False)
basics = basics[~is_documentary]

**Update AKAs dataframe to eliminate the movies eliminated from basics**

akas = akas.loc[akas['titleId'].isin(basics['tconst'])]

**Keep only movies released in the US**

usAKAs = akas.loc[akas['region'] == 'US']

basics = basics.loc[basics['tconst'].isin(usAKAs['titleId'])]

ratings = ratings.loc[ratings['tconst'].isin(basics['tconst'])]

basics.shape

ratings.shape

import os
os.makedirs('Data/',exist_ok=True) 
# Confirm folder created
os.listdir("Data/")

## Save dataframes to Compressed .csv.gz files
basics.to_csv("Data/basics.csv.gz",compression='gzip',index=False)
ratings.to_csv("Data/ratings.csv.gz",compression='gzip',index=False)
usAKAs.to_csv("Data/akas.csv.gz",compression='gzip',index=False)

# Open saved file and preview
basics = pd.read_csv("Data/basics.csv.gz", low_memory = False)
basics.head()
