==> 03_Consolidate_Data_from_API_Extract.ipynb <==
import pandas as pd
import os

FOLDER = 'TMDb_Data/'
file_list = sorted(os.listdir(FOLDER))
file_list

import glob
## Make a filepath query
q = FOLDER+"final*.csv.gz"
print(q)
file_list = sorted(glob.glob(q))
file_list

file_list[:2]

# proof of concept, years 2000 and 2001
df_2yrs = pd.concat([pd.read_csv(file) for file in file_list[:2]])
df_2yrs.reset_index(inplace=True, drop=True)
df_2yrs

df_2yrs.to_csv("Final_Data/tmdb_2000_2001_results_combined.csv.gz", compression="gzip", index=False)
# this was requested as a proof of concept

# now extract all data
df = pd.concat([pd.read_csv(file) for file in file_list])
df.reset_index(inplace=True, drop=True)
df

**Cleaning Data** 
* First to remove all movies with an imdb_id of 0 (this is an artifact of how we set up API call. If the file did not exist yet, we made a new file with one movie with an id of 0. Nevertheless, this can now be removed.)

df = df[df['imdb_id'] != '0']
df.reset_index(inplace=True, drop=True)
df

df.duplicated().sum()

df.to_csv("Final_Data/tmdb_ALL_results_combined.csv.gz", compression="gzip", index=False)
