==> 05_ETL_to_mySQLdb.ipynb <==
# Specifications - Database
* Stakeholder wants you to take the data you have been cleaning and collecting in Parts 1 & 2 of the project, and wants you to create a MySQL database for them.

* I will normalize the tables as best I can before adding them to the new database.

    * Note: an important exception to their request is that they would like you to keep all of the data from the TMDB API in 1 table together (even though it will not be perfectly normalized).  
    * You only need to keep the imdb_id, revenue, budget, and certification columns

## Genre
### Normalization
* Convert the single string of genres from title basics into 2 new tables.
    * title_genres: with the columns:
        - tconst
        - genre_id
    * genres:
        - genre_id
        - genre_name

#imports
import pandas as pd
import numpy as np
import pymysql
pymysql.install_as_MySQLdb()
from sqlalchemy import create_engine
from sqlalchemy_utils import create_database, database_exists
connection_str = "mysql+pymysql://root:root@localhost/Movies"

engine = create_engine(connection_str)
if database_exists(connection_str) == False: create_database(connection_str)
else: print('The database already exists.')

df = pd.read_csv('Data/basics.csv.gz', low_memory = False)

The genres column needs to be split into separate genres:

df['genres']

df['genre-split'] = df['genres'].str.split(',')
df

exploded_genres = df.explode('genre-split')
exploded_genres[['tconst', 'primaryTitle', 'genre-split']]

unique_genres = sorted(exploded_genres['genre-split'].dropna().unique())
unique_genres

title_genres = exploded_genres[['tconst', 'genre-split']].copy()
title_genres

## Making the genre mapper dictionary
genre_ints = range(len(unique_genres))
genre_map = dict(zip(unique_genres, genre_ints))
genre_map

## make new integer genre_id and drop string genres
title_genres['genre_id'] = title_genres['genre-split'].map(genre_map)
title_genres = title_genres.drop(columns='genre-split')
title_genres

genre_lookup = pd.DataFrame({'Genre_Name': genre_map.keys(), 
                            'Genre_ID': genre_map.values()})
genre_lookup.head()

title_genres

### Export Title_Genres Table to MySQL

title_genres.dtypes

from sqlalchemy.types import *
## Calculate max string lengths for object columns
key_len = title_genres['tconst'].fillna('').map(len).max()

## Create a schema dictonary using Sqlalchemy datatype objects
title_genre_schema = {
    "tconst": String(key_len+1), 
    'genre_id':Integer()}

# Save to sql with dtype and index=False
title_genres.to_sql('title_genres',engine,dtype=title_genre_schema,if_exists='replace',index=False)

Note: `title_genres` is a connector table and does not have a primary key.

q = """
SELECT *
FROM title_genres
LIMIT 5
"""
pd.read_sql(q, engine)

### Export Genres Table to MySQL

genre_lookup.dtypes

genre_name_len = genre_lookup['Genre_Name'].fillna('').map(len).max()
genre_lookup_schema = {
    "Genre_Name": Text(genre_name_len+1),
    'Genre_ID':Integer()}

# Save to sql with dtype and index=False
genre_lookup.to_sql('genres',engine,dtype=genre_lookup_schema,if_exists='replace',index=False)

engine.execute('ALTER TABLE genres ADD PRIMARY KEY (`Genre_ID`);')

q = """
SELECT *
FROM genres
LIMIT 5
"""
pd.read_sql(q, engine)

## Basics

Now discard unnecessary information from basics table:
* "original_title" (we will use the primary title column instead)
* "isAdult" ("Adult" will show up in the genres so this is redundant information).
* "titleType" (every row will be a movie).
* "genres" and other variants of genre (genre is now represented in the 2 new tables described above.

df

df = df.drop(columns = ['originalTitle','isAdult', 'titleType', 'genres', 'genre-split'])
df

df.dtypes

title_len = df['primaryTitle'].fillna('').map(len).max()
basics_schema = {
    "tconst": String(key_len+1), #key_len defined above
    "primaryTitle": Text(title_len+1),
    'startYear':Integer(),
    'endYear':Float(),
    'runtimeMinutes':Integer()}

df.to_sql('title_basics',engine,dtype=basics_schema,if_exists='replace',index=False)

engine.execute('ALTER TABLE title_basics ADD PRIMARY KEY (`tconst`);')

q = """
SELECT *
FROM title_basics
LIMIT 5
"""
pd.read_sql(q, engine)

## Ratings

ratings = pd.read_csv('Data/ratings.csv.gz', low_memory = False)

ratings.dtypes

ratings_schema = {
    "tconst": String(key_len+1), #key_len defined above
    "averageRating": Float(),
    'numVotes':Integer()}

ratings.to_sql('title_ratings',engine,dtype=ratings_schema,if_exists='replace',index=False)

engine.execute('ALTER TABLE title_ratings ADD PRIMARY KEY (`tconst`);')

q = """
SELECT *
FROM title_ratings
LIMIT 5
"""
pd.read_sql(q, engine)

## TMDB Data

tmdb_data = pd.read_csv('Final_Data/tmdb_ALL_results_combined.csv.gz', low_memory = False)
tmdb_data = tmdb_data[['imdb_id', 'revenue', 'budget', 'certification']]
tmdb_data

tmdb_data['certification'].value_counts(dropna = False)

tmdb_data['certification'] = tmdb_data['certification'].str.strip()
tmdb_data['certification'].value_counts(dropna=False)

## fix certification col
repl_cert = {'UR':'NR',
             'Not Rated':'NR',
             'Unrated':'NR',
             '-':'NR',
             '10':np.nan,
             'ScreamFest Horror Film Festival':np.nan}
tmdb_data['certification'] = tmdb_data['certification'].replace(repl_cert)
tmdb_data['certification'].value_counts(dropna=False)

tmdb_data.dtypes

certification_len = tmdb_data['certification'].fillna('').map(len).max()
tmdb_schema = {
    "imdb_id": String(key_len+1), #key_len defined above
    "revenue": Float(),
    'budget':Float(),
    'certification': Text(certification_len+1)}

tmdb_data.to_sql('tmdb_data',engine,dtype=tmdb_schema,if_exists='replace',index=False)

engine.execute('ALTER TABLE tmdb_data ADD PRIMARY KEY (`imdb_id`);')

q = """
SELECT *
FROM tmdb_data
LIMIT 5
"""
pd.read_sql(q, engine)

## Testing Database

q = """SHOW TABLES;"""
pd.read_sql(q, engine)
