==> 07_Predicting Movie Success.ipynb <==
# Predicting Movie Success

*For part 5 of the project, you will be applying your Linear Regression knowledge to create a Linear Regression model that meets the 4 assumptions of Linear Regression. Additionally, you will then use your model's results to make 3 recommendations for your stakeholder on how to create a successful movie.*

* While this part of the project is optional and not graded, it would be a fantastic piece for your portfolio!

    - Note: predicting revenue is the most obvious target for describing "success", the average rating is also a logical choice that meets your stakeholder's request.
    
## Specifications
* You should test your linear regression model for all 4 assumptions.
* You should show an iterative modeling process (keep multiple models in the final notebook, showing the improvement in performance/assumptions.
* You should use the coefficients from your model to create 3 recommendations on how to make a successful movie.

## Deliverables
* You should use the same project repository you have been using for Parts 1-5 (for your portfolio).
* Create a new notebook in your project repository just for the hypothesis testing (like "Part 5 - Predictive Models.ipynb" or "Part 5 - Linear Regression Models.ipynb" )
* You should make a short presentation summarizing your findings and recommendations, and include visualizations to support each recommendation.


Please submit the link to your GitHub repository for this assignment and reach out to your instructor on Discord or via email if you would like to get feedback on your project.
